Prepared conv1d layers for tfjs.
I work towards adopting the module to allow for a seemless integration
with `tensoflow.js`.  My usecase targets Conv1d layers, which is why
changes to convert_conv are only partial.  There are two main changes:

* pytorch uses 'NCW' which needs to be converted to 'NWC' for
tensorflow.  However, we want to retain the cool names of our graph's
operators, which is why we need to create a temporary new one that will
hold the ops with changed order.

* PaddingLayers are not supported by tfjs as of now.  We therefore allow
for a 'valid' mode conv1d layer where the padding helper-layer is
omitted.  We also allow to provide the data-format in `params`.

* We need to avoid the long random-number names b/c these aren't
supported by tfjs.